move db to manga db
