+++ tanoshi/src/context.rs
@@ -1,14 +1,14 @@
 use crate::catalogue::*;
 use crate::extension::Extensions;
-use crate::db::Db;
+use crate::db::MangaDatabase;
 
 pub struct GlobalContext {
-    pub db: Db,
+    pub mangadb: MangaDatabase,
     pub extensions: Extensions,
 }
 
 impl GlobalContext {
-    pub fn new(db: Db, extensions: Extensions) -> Self {
-        Self { db, extensions }
+    pub fn new(mangadb: MangaDatabase, extensions: Extensions) -> Self {
+        Self { mangadb, extensions }
     }
 }
